+++ src/Api/Providers/News.php
@@ -9,15 +9,31 @@ class News extends Provider
     /**
      * @var array
      */
-    protected $loginRequiredFor = ['last'];
+    protected $loginRequiredFor = ['last', 'all'];
 
     /**
      * Get user's latest news array.
+     * @deprecated 4.11.0
      *
-     * @return array|bool
+     * @param $limit
+     * @return mixed
      */
-    public function last()
+    public function last($limit = 0)
     {
-        return $this->execGetRequest(['allow_state' => true], UrlBuilder::RESOURCE_GET_LATEST_NEWS);
+        return $this->all($limit);
+    }
+
+    /**
+     * @param int $limit
+     * @return mixed
+     */
+    public function all($limit = 0)
+    {
+        $params = [
+            'data' => ['allow_stale' => true],
+            'url'  => UrlBuilder::RESOURCE_GET_LATEST_NEWS
+        ];
+
+        return $this->getPaginatedResponse($params, $limit);
     }
 }