+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [4.11.0] - ?
+### Added:
+ - User *changePassword*, *deactivate* methods.
+ - AuthRequired exception.
+ - Pins *getRelatedPins* method.
+ - Pinners *likes* method.
+ - News *all* method.
+ - Boards *getTitleSuggestionsFor* method.
+ 
+### Changed:
+ - User *login* method does not throw an exception, now it returns *false* on fail.
+ - Bot *getLastError* method returns string, not array.
+ - News *last* method is deprecated.
+
 ## [4.10.1] - 2016-09-13
 ### Fixed:
  - Checks in *getUserName* and *isBanned* methods